refactor: change components file extension to tsx
