+++ assets/Components/List/DirectoryItem.tsx
