+++ assets/Components/Thumbnail.tsx
